Fix typo in cMapSender constructor. Avoid some std::vector copy in cUnitsData::initializeClanUnitData. Replace some "double" function by "float" function. Some clean up: - Add const - truncate some separator comment to 80 columns. - use some std::min
git-svn-id: svn://maxr.org/MAX/trunk@3235 c3bd5e88-ca0b-48cf-b972-da51c77c4c1b

+++ hud.cpp
@@ -1045,7 +1045,7 @@ SDL_Surface* cGameGUI::generateMiniMapSurface()
 		//to make sure, that every unit is diplayed and has the same size on the minimap.
 
 		//the size of the rect, that is drawn for each unit
-		int size = (int) ceil ((float) MINIMAP_SIZE * zoomFactor / map->getSize());
+		int size = (int) ceilf ((float) MINIMAP_SIZE * zoomFactor / map->getSize());
 		size = std::max (size, 2);
 		SDL_Rect rect;
 		rect.h = size;
